+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## [Unreleased]
 ### Added
+- File Upload
 
 ### Changed
+- fixed: source was re-authenticating every request (DateTime math was incorrect)
 
 ### Removed